Fix clang-tidy warning

+++ src/server.c
@@ -61,7 +61,7 @@ int http_server_init(HttpServer *server, const char *port) {
 }
 
 static int send_all(int fd, const char *buf, size_t len) {
-  for (int sent = -1; len; buf += sent, len -= sent) {
+  for (ssize_t sent = -1; len; buf += sent, len -= sent) {
     sent = send(fd, buf, len, 0);
     if (sent == -1) {
       perror("send");
